working image generation

+++ pitch_extraction.py
@@ -238,6 +238,6 @@ class FeatureExtractor:
         ) + 1
         f0_mel[f0_mel <= 1] = 1
         f0_mel[f0_mel > 255] = 255
-        f0_coarse = np.rint(f0_mel).astype(np.int)
+        f0_coarse = np.rint(f0_mel).astype(np.int16)
 
         return f0_coarse, f0  # 1-0
